Fix unresolvable development dependency tree
The number of combinations of dependencies was so large that Bundler was taking
a very long time (forever?) to resolve the dependencies. Pinning the dev
dependencies doesn't affect users of the gem.

Also require Rails 3.2 and above - not strictly required in this change, but
all of our apps are on 3.2 and above anyway.