Add :before_content content_for block to layout template
This adds a little more flexibility in the layout you can acheive from
a consuming app's layout.

The use case here was the need to have a flash message rendered above
content in the layout. As layouts are rendered after templates, content
can only be appended to the end of a content_for block from a layout,
hence the need for a distinct :before_content content_for block.