Add a radio toggle
Create a radio toggle so that content can be conditionally revealed
and forms simplified.

The JS is based on GOV.UK elements patterns, setting aria controls and
aria hidden attributes dynamically. A radio and a toggle target are
linked with an ID.

The module checks the current state of the form and then hides or shows
content based on that.

The radio example is a little more convoluted than the checkbox
equivalent. It needs to listen for changes to all radio elements with
the same name – when one radio is selected there is no corresponding
change event on the others.