Enable sendBeacon for all analytics events
In some situations we will be tracking events that are triggered just
before the page unloads, in these situations we will need sendBeacon.

Enable this for all events so the implementor doesn’t need to consider
the nature of the event they’re creating. There’s no known downside to
this approach.