Add TrackButtonPress module
This module is intended as an easy way to track individual button/link
clicks as events in Google Analytics. Groups of buttons may end up
redirecting to the same place, meaning they’re not trackable via URLs
in GA. This will use the sendBeacon functionality to fire off an event
once the button has been clicked and during navigation.

The module should be added to the container of the buttons, where the
GA action can be changed from the default of `button-pressed` using the
`data-track-action` attribute. The label defaults to the link/button
text, but can be overridden with `data-track-label`.