Check module HTML explicitly in tests
The tests contain a deficiency that was revealed when porting them to
govuk_frontend_toolkit. The existing module assertions use `toMatch` on
HTML objects which gives a false positive.

Instead assert that the HTML we expect is within the jQuery object.

This unveiled a failing test which has also been fixed. The `push`
method for when a container is also a module should actually be `add`.