Add configuration
This commit adds a `Config` module we can use to provide application
configs. It will be used to configure certain layout elements that
should be opt-in, like a flash message (which is not compatible with
some apps).

To demonstrate usage we can now configure the site title in the config
block, instead of using the `content_for :page_title` block.