Add optional signout element to nav bar
`config.show_signout` will make the current username and signout link
appear in the top right corner. This is a common interface element,
used in many apps like collections-publisher and search-admin.

It is still opt-in to prevent apps from having to upgrade.

To make the dummy app work, we've had to add a fake `current_user`
method and a fake `Plek` class.

The method is "enabled" in the readme config, so that new apps use this
config.