Minimum requirements for Rails 5 dependencies
This needed bumping Ruby to 2.2.2 for Rack 2 and jquery-rails /
sass-rails for updates to railties.

The dummy app was basically Rails 5 compatible, save for the
`active_resource` mentions (now obviously defunct), and a few minor
breaking changes. As the assets are now fingerprinted I made the spec
expectations more permissive to make them pass.