Adds disable_google_analytics setting to config
Enables analytics to be disabled using a config setting. The gem currently assumes that all projects
created with the govuk template want to enable google analytics. This setting will not change current
projects using the gem, but will make it less opinionated about using analytics.

This will avoid the potential surprise of an application crashing when being deployed to production
because it is expecting google analytics credentials.