Switch to a more flexible method signature
[Trello card](https://trello.com/c/BW5fA4ar/141-normalizes-ga-tracking-for-admin-gem)

This is a breaking change to the API.

`GOVUKAdmin.trackEvent(options)` will replace `GOVUKAdmin.trackEvent(action, label, value, category)`
Category is typically the first parameter, and if using the Google docs as reference is required. The original method call we were using was not useful.

This PR came about as a result of the changes made here: https://github.com/alphagov/govuk_admin_template/pull/145