Remove Rails 5.1 asset fallback warning in test
Rails / Sprockets now spews deprecation messages when failing to
resolve the favicon asset for the `test` environment unless the
`skip_pipeline: true` option is passed - as follows:

```
DEPRECATION WARNING: The asset "govuk_admin_template/favicon-test.png"
is not present in the asset pipeline.Falling back to an asset that may
be in the public folder.  This behavior is deprecated and will be
removed.  To bypass the asset pipeline and preserve this behavior, use
the `skip_pipeline: true` option.
```

I've copied the development favicon to the test favicon and added this
to the list of assets to precompile thus removing any warnings.