Add assets:clobber placeholder rake task
Like the other assets:* tasks this is just an empty placeholder that does
nothing.  They're only here to let us run the `precompileAssets` step in
the standard `buildProject` we use on Jenkins.  We may wish to copy what
govuk_publishing_components does which is have these assets:* tasks
delegate down to the dummy_app:assets:* tasks so that they actually try
to do something.