Allow setting a custom class on the root HTML tag
This change allows the insertion of a custom class on the root
`<html>` tag in the Admin Template.

This is useful when making more drastic styling changes to the template.
The class is set at such a high level in the DOM because this allows:

- changing of styles on the `html` tag itself (ie `overflow-y`)
- arbitrary restyling of anything else in the DOM

This change is primarily to allow a side-by-side layout in the Audit
Tool.