Add dependabot configuration
We've decided to keep the dependabot config quite lightweight – without specifying which gems to 'allow', as is commonly done in other GOV.UK projects.

We did this because this project's gemspec doesn't specify many dependencies, so allowing all updates is unlikely to cause a dependabot-storm. For some gems, we don't expect (m)any updates anyway – e.g. bootstrap-sass and jquery-rails.

We can always change this in the future if it gets too noisy.