Require Plek in the Dummy App
The dummy app located in `spec/dummy` couldn't be started because the Plek gem wasn't being included in `application.rb`.

Whilst this isn't needed for tests to run, it's useful to be able to fire up the dummy app in dev mode and view it in your browser.

To run the app:

```
cd spec/dummy
rails s
```

and it'll be available in your browser at:
http://localhost:3000