Fix broken copyright link in footer
The "Crown Copyright" link in the page footer was broken. It looks like the National Archives website, which hosts this page, has moved the it to a different URL.

I've replaced the old broken link with its new equivalent, taken from the footer of the current https://gov.uk homepage.